Updates due to bad handling of last merge

+++ star/star_formation.f90
@@ -65,7 +65,7 @@ subroutine star_formation(r,g,m,s,ilevel,mstar_loc)
   integer(kind=8),dimension(0:g%ncpu)::nsite_cum,nstar_cum
   integer,dimension(1:g%ncpu)::nsite_cpu,nstar_cpu
   integer::i,ind,igrid,idim,icpu,ngrid,nleaf,nsite,nstar,nstar_loc
-  real(dp)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v,erfc
+  real(kind=8)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v,erfc
   real(kind=8)::dx,vol,factG,n_star,nCOM,d,d0,mstar,dstar,t_ff,mcell,mgas,mask,PoissMean,Rand
   real(kind=8)::sfr_ff,t_dyn,p,cs2,sigma2,alpha_vir,sigs,scrit,Mach2,b_turb
 #if NENER>0
